src/news/api.py
from ninja import NinjaAPI
from django.shortcuts import render

api = NinjaAPI()

@api.get("")
def home(request):
    return render(request, "news/home.html")

@api.get("/latest")
def latest(request):
    """
    Render a Datatable (tablerio) of all the videos
    """
    return 0
