requirements.txt
Django==5.0.3
django-ninja
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.3.0

psycopg==3.1.18

redis==5.0.3
celery==5.3.6

flake8==7.0.0
isort==5.13.2
black==24.2.0

python-dotenv
google-api-python-client
protobuf
